Fixed order of TTFT,TPOT,E2Es,CompletionTimes
