+++ sim/metrics.go
@@ -3,14 +3,8 @@
 package sim
 
 import (
-	"bufio"
 	"fmt"
-	"math"
-	"os"
-	"sort"
 	"time"
-
-	"github.com/sirupsen/logrus"
 )
 
 // Metrics aggregates statistics about the simulation
@@ -29,105 +23,13 @@ type Metrics struct {
 	TTFTSum int64 // Total time-to-first-token sum (in ticks)
 	TPOTSum int64 // Total TPOT sum across requests (in ticks)
 
-	RequestTTFTs           []float64 // list of all requests' TTFT
-	RequestTPOTs           []float64 // list of all requests' TPOT
-	RequestE2Es            []float64 // list of all requests' latencies
-	RequestCompletionTimes []float64 // list of all requests' completion times in ticks
+	RequestTTFTs           map[string]float64 // list of all requests' TTFT
+	RequestTPOTs           map[string]float64 // list of all requests' TPOT
+	RequestE2Es            map[string]float64 // list of all requests' latencies
+	RequestCompletionTimes map[string]float64 // list of all requests' completion times in ticks
 
 	NumWaitQRequests        []int // number of requests in waitQ over different steps
 	NumRunningBatchRequests []int // number of request in runningBatch over different steps
-}
-
-// Bin represents a single completion time integral bin with its integer key and count.
-type Bin struct {
-	Key   int
-	Count int
-}
-
-// CalculatePercentile is a util function that calculates the p-th percentile of a data list
-func CalculatePercentile(data []float64, p float64) float64 {
-	n := len(data)
-
-	sortedData := make([]float64, n)
-	copy(sortedData, data)
-
-	sort.Float64s(sortedData)
-
-	rank := p / 100.0 * float64(n-1)
-	lowerIdx := int(math.Floor(rank))
-	upperIdx := int(math.Ceil(rank))
-
-	if lowerIdx == upperIdx {
-		return sortedData[lowerIdx]
-	} else {
-		lowerVal := sortedData[lowerIdx]
-		upperVal := sortedData[upperIdx]
-		if upperIdx >= n {
-			return sortedData[n-1]
-		}
-		return lowerVal + (upperVal-lowerVal)*(rank-float64(lowerIdx))
-	}
-}
-
-func CalculateBinnedThroughput(data []float64) float64 {
-	tempBins := make(map[int]int)
-	maxBinKeyActual := 0
-	totalCompletionCounts := 0
-
-	if len(data) == 0 {
-		return 0
-	}
-
-	for _, val := range data {
-		binKey := int(val)
-		tempBins[binKey]++
-		maxBinKeyActual = max(maxBinKeyActual, binKey)
-	}
-
-	for i := 0; i <= maxBinKeyActual; i++ {
-		if _, ok := tempBins[i]; !ok {
-			tempBins[i] = 0
-		}
-	}
-
-	for _, count := range tempBins {
-		totalCompletionCounts += count
-	}
-
-	meanCompletionCount := float64(totalCompletionCounts) / float64(len(tempBins))
-
-	return meanCompletionCount
-}
-
-func (m *Metrics) SavetoFile(data []int, fileName string) {
-	file, err := os.OpenFile(fileName, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0777)
-	if err != nil {
-		logrus.Fatalf("Error creating file %s: %v\n", fileName, err)
-		return
-	}
-	defer func() {
-		if closeErr := file.Close(); closeErr != nil {
-			logrus.Fatalf("Error closing file %s: %v\n", fileName, closeErr)
-		}
-	}()
-
-	writer := bufio.NewWriter(file)
-
-	defer func() {
-		if flushErr := writer.Flush(); flushErr != nil {
-			logrus.Fatalf("Error flushing writer for file %s: %v\n", fileName, flushErr)
-		}
-	}()
-
-	for _, f := range data {
-		_, writeErr := fmt.Fprint(writer, f, ", ")
-		if writeErr != nil {
-			logrus.Fatalf("Error writing int %d to file: %v\n", f, writeErr)
-			return // Stop writing on first error
-		}
-	}
-
-	logrus.Debugf("Successfully wrote to '%s'\n", fileName)
 }
 
 // Print displays aggregated metrics at the end of the simulation.
@@ -142,16 +44,20 @@ func (m *Metrics) Print(horizon int64, totalBlocks int, startTime time.Time) {
 	fmt.Printf("vLLM estimated Duration(s): %d\n", m.SimEndedTime/1e6)
 	if m.CompletedRequests > 0 {
 		avgTTFT := float64(m.TTFTSum) / float64(m.CompletedRequests)
-		medianTTFT := CalculatePercentile(m.RequestTTFTs, 50)
-		p99TTFT := CalculatePercentile(m.RequestTTFTs, 99)
+		sortedTTFTs := SortRequestMetrics(m.RequestTTFTs)
+		sortedTPOTs := SortRequestMetrics(m.RequestTPOTs)
+		sortedE2Es := SortRequestMetrics(m.RequestE2Es)
+		sortedCompletionTimes := SortRequestMetrics(m.RequestCompletionTimes)
+		medianTTFT := CalculatePercentile(sortedTTFTs, 50)
+		p99TTFT := CalculatePercentile(sortedTTFTs, 99)
 		avgTPOT := float64(m.TPOTSum) / float64(m.TotalOutputTokens)
-		medianTPOT := CalculatePercentile(m.RequestTPOTs, 50)
-		p99TPOT := CalculatePercentile(m.RequestTPOTs, 99)
-		perSecThroughput := CalculateBinnedThroughput(m.RequestCompletionTimes)
+		medianTPOT := CalculatePercentile(sortedTPOTs, 50)
+		p99TPOT := CalculatePercentile(sortedTPOTs, 99)
+		perSecThroughput := CalculateBinnedThroughput(sortedCompletionTimes)
 
 		fmt.Printf("Request throughput (req/s):  : %.3f\n", perSecThroughput)
 		fmt.Printf("TTFTs             :[")
-		for _, ttft := range m.RequestTTFTs {
+		for _, ttft := range sortedTTFTs {
 			fmt.Printf("%.6f, ", ttft/1000)
 		}
 		fmt.Printf("]\n")
@@ -159,7 +65,7 @@ func (m *Metrics) Print(horizon int64, totalBlocks int, startTime time.Time) {
 		fmt.Printf("Median TTFT(ms)   : %.3f\n", medianTTFT)
 		fmt.Printf("P99 TTFT(ms)      : %.3f\n", p99TTFT)
 		fmt.Printf("TPOTs             : [")
-		for _, tpot := range m.RequestTPOTs {
+		for _, tpot := range sortedTPOTs {
 			fmt.Printf("%.6f, ", tpot/1000)
 		}
 		fmt.Printf("]\n")
@@ -167,8 +73,8 @@ func (m *Metrics) Print(horizon int64, totalBlocks int, startTime time.Time) {
 		fmt.Printf("Median TPOT(ms)   : %.3f\n", medianTPOT)
 		fmt.Printf("P99 TPOT(ms)      : %.3f\n", p99TPOT)
 		fmt.Printf("E2Es             : [")
-		for _, e2e := range m.RequestE2Es {
-			fmt.Printf("%.3f, ", e2e/1000)
+		for _, e2e := range sortedE2Es {
+			fmt.Printf("%.6f, ", e2e/1000)
 		}
 		fmt.Printf("]\n")
 		fmt.Printf("Avg KV Blocks Usage : %.3f\n", float64(m.KVBlocksUsed)/float64(m.SimEndedTime))
